Move mutation observer code back to CStoreBase
